Move Button component to a separated file
